+++ core/worker/config.go
@@ -2,6 +2,7 @@ package worker
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
@@ -12,12 +13,15 @@ func init() {
 	viper.BindEnv(workerSlotsViperKey, workerSlotsEnv)
 	viper.SetDefault(workerPartitionsViperKey, workerPartitionsDefault)
 	viper.BindEnv(workerPartitionsViperKey, workerPartitionsEnv)
+	viper.SetDefault(workerTimeoutViperKey, workerTimeoutDefault)
+	viper.BindEnv(workerTimeoutViperKey, workerTimeoutEnv)
 }
 
 // Config is worker configuration
 type Config struct {
 	Slots      int64
 	Partitions int64
+	Timeout    time.Duration
 }
 
 // Validate ensure configuration is valid
@@ -30,6 +34,10 @@ func (c *Config) Validate() error {
 		return fmt.Errorf("At least one partition is required")
 	}
 
+	if c.Timeout == 0 {
+		return fmt.Errorf("Timeout must be positive")
+	}
+
 	return nil
 }
 
@@ -38,6 +46,7 @@ func NewConfig() Config {
 	config := Config{}
 	config.Slots = viper.GetInt64("worker.slots")
 	config.Partitions = viper.GetInt64("worker.partitions")
+	config.Timeout = viper.GetDuration("worker.timeout")
 
 	return config
 }
@@ -46,6 +55,7 @@ func NewConfig() Config {
 func InitFlags(f *pflag.FlagSet) {
 	Slots(f)
 	Partitions(f)
+	Timeout(f)
 }
 
 var (
@@ -76,4 +86,19 @@ func Partitions(f *pflag.FlagSet) {
 Environment variable: %q`, workerPartitionsEnv)
 	f.Uint(workerPartitionsFlag, workerPartitionsDefault, desc)
 	viper.BindPFlag(workerPartitionsViperKey, f.Lookup(workerPartitionsFlag))
+}
+
+var (
+	workerTimeoutFlag     = "worker-timeout"
+	workerTimeoutViperKey = "worker.timeout"
+	workerTimeoutDefault  = 60 * time.Second
+	workerTimeoutEnv      = "WORKER_TIMEOUT"
+)
+
+// Timeout register flag for Kafka server addresses
+func Timeout(f *pflag.FlagSet) {
+	desc := fmt.Sprintf(`Maw time to handle a message
+Environment variable: %q`, workerTimeoutEnv)
+	f.Duration(workerTimeoutFlag, workerTimeoutDefault, desc)
+	viper.BindPFlag(workerTimeoutViperKey, f.Lookup(workerTimeoutFlag))
 }